Store lock as a function attribute
Rather than store the lock in a dictionary, which at best uses the
Global Interpreter Lock to get a value from a dictionary (and dictionary
reads are not guaranteed thread safe), store the lock as an attribute on
the function, which should never have access issues. It also removes a
"global" parameter from the object.